+++ scripts/DEM_Analysis_References/2_Blocks/200_SW_Problem.py
@@ -1,60 +1,47 @@
-from compas_model.materials import Concrete
+import os
 
-from compas_dem.models import BlockModel
+import compas
+
 from compas_dem.problem import Problem
-from compas_dem.problem import Solver
-from compas_dem.templates import ArchTemplate
 from compas_dem.viewer import DEMViewer
 
 # =============================================================================
-# Template
+# Load model
 # =============================================================================
 
-template = ArchTemplate(rise=3, span=10, thickness=0.5, depth=0.5, n=50)
+HERE = os.path.dirname(__file__)
+model = compas.json_load(
+    os.path.join(HERE, "DEM_model.json"),
+)
 
 # =============================================================================
-# Model
-# =============================================================================
-
-model = BlockModel.from_template(template)
-
-# =============================================================================
-# Interactions
-# =============================================================================
-
-model.compute_contacts(tolerance=0.001)
-
-# =============================================================================
-# Supports
-# =============================================================================
-
-for node in model.graph.nodes_where(degree=1):
-    model.graph.node_element(node).is_support = True  # type: ignore
-
-# ============================================================================
-# Material
-# ============================================================================
-
-conc: Concrete = Concrete.from_strength_class("C30")
-model.add_material(conc)
-model.assign_material(conc, elements=list(model.elements()))
-
-# =============================================================================
-# Problem
+# Create Problem
 # =============================================================================
 
 problem = Problem(model)
-problem.add_contact_model("MohrCoulomb", mu=0.5, c=0.0)
+
+# =============================================================================
+# Add supports
+# =============================================================================
+
 problem.add_supports_from_model()
 
-lmgc90: Solver = Solver.LMGC90(dt=0.01, n_steps=100)
-problem.solve(lmgc90)
-
 # =============================================================================
-# Viz
+# Add contact properties
 # =============================================================================
 
-viewer = DEMViewer(model)
+problem.add_contact_model("MohrCoulomb", mu=0.5)
 
-viewer.add_solution()
+# =============================================================================
+# Save problem
+# =============================================================================
+
+HERE = os.path.dirname(__file__)
+compas.json_dump(problem, os.path.join(HERE, "DEM_problem.json"))
+# =============================================================================
+# Visualize problem
+# =============================================================================
+
+viewer = DEMViewer(problem.model)
+viewer.setup()
 viewer.show()