+++ src/compas_dem/material/stone.py
@@ -1,7 +1,6 @@
 from typing import Literal
 from typing import Optional
 
-from compas_model.materials import Concrete
 from compas_model.materials import Material
 from compas_model.materials.errors import PropertyNotDefined